Move all value provider decorators to one package
